requirements.txt
discord.py
flask
python-dotenv
psutil
nekos.py
owotext